+++ src/types.ts
@@ -1,6 +1,7 @@
 import { ProxyDiContainer } from './ProxyDiContainer';
 
 export type DependencyId = string | symbol;
+export type AliasId = string | symbol;
 
 export type DependencyClass<T> = new (...args: any[]) => T;
 
@@ -17,13 +18,9 @@ export type SingleInjection = {
     property: string | symbol;
     dependencyId: DependencyId;
     set: Setter;
-    //isAll?: false;
 };
 
 export type AllInjection = SingleInjection & {
-    // property: string | symbol;
-    // dependencyId: DependencyId;
-    // set: Setter;
     isAll: true;
     scope: ResolveScope;
 };
@@ -45,7 +42,10 @@ export type IProxyDiContainer = {
 
     parent?: IProxyDiContainer;
 
-    isKnown: (dependencyId: DependencyId | DependencyClass<any>, scope?: ResolveScope) => boolean;
+    isKnown: (
+        dependencyId: DependencyId | DependencyClass<any>,
+        scope?: ResolveScope
+    ) => boolean;
 
     injectDependenciesTo: (dependency: any) => void;
 
@@ -118,7 +118,7 @@ export type ContainerizedDependency = Dependency & {
 
 export type ContainerSettings = {
     allowRewriteDependencies?: boolean;
-    resolveInContainerContext?: boolean;
+    resolveInContainerContext?: boolean; // TODO: Change to contextResolve()
 };
 
 export const IS_INJECTION_PROXY = Symbol('isInjectionProxy');